Rename Inventory to StorageLocation
This is a *huge* change. Renamed the `Inventory` model to
`StorageLocation` because the type of information being captured in that
model better reflects something that is a place, rather than a concept
for holding items in inventory. If anything, the `#inventory_items`
assocation represents the actual inventory, so that might be worth
renaming...again.

This touched a lot of stuff because the concept of inventory is related
to donations, distributions, and inventory items.

All tests are passing.

closes #41