Wires up transfering inventory betweeen locations
Sets up the basic view capability and controller logic for creating a
transfer so inventory can be exchanged between storage facilities.

When the user changes their source location for a transfer, the list of
items that can be transferred from that source is now dynamically
generated to reflect only the items available at that source storage
location.

closes #25