Consistent form layout and responsiveness
Converted all of the forms to use a vertical layout (where the label
sites above the input control). Also adjusts the responsiveness of the
forms to be centered for all screen resolutions until the screen gets
too small. In that case the controls will fit the width of the screen.