Fixes margins for some tables and forms
Makes the content for the right-most “action buttons” cell in a row fill the entire cell and removes the gap in the row border
Restores form container width to parent container width instead of full browser width. The left sidebar reduces the available browser width for content but the forms were wrapped in a ‘container’ class that uses the entire browser width.