Sets selected items when updating inventory lists
Fixes a bug when creating distributions and transfers that fails
validation (i.e., distributing more inventory than is available).
When the validation fails and kicks back to the new
distribution/transfer form, the selected inventory item for each
line item defaults back to the first option in the dropdown.

This is because Rails first populates the dropdown list
with the entire list of inventory items, and even properly selects the
correct item for each line item. However, after the page is loaded a
JavaScript function kicks in and replaces each of the inventory
dropdowns with the available inventory for the selected storage
facility. When this happens the selected option is blown away.

This fix properly determines what the selected `<option>` tag should be
and marks it when regenerating the dropdown lists for each line item.
It's implemented so that it uses the same JavaScript code to support
distributions and transfers.

Fixes #300