Move controller-helper items collection to a query object
There are two tabs on http://localhost:3000/pdx_bank/items that need a special
query.

In the old code, the query lived in the controller (as a pseudo-helper),
then got massaged in place in the view to aggregate count by location.

This commit does the following:
- Move the query into a query object
- Aggregated count also gets moved into a query object
  - Don't try to be smart, make the code readable: call the first query object
    and iterate over it to aggregate count by location
- Remap second-tab view's instance variable to use the new query object
- Rewrite third-tab view to use the newly returned hash (this is the one that
  used to massage in place)
- Delete the old code

Note: there's minimal refactoring in this commit. Make sure it works, then
consider whether to make it cleaner.