Convert authorization feature test to system test
This fixes Issue #826

Also I changed the language of the test from "It should.." to "It
does..". I think it makes it more readable. However, now is not
consistent with the rest of the tests, so I'm happy to change it back if
you think it's better in the old way.