Switch nested fields to simple_form
Fixes #1103

This also adds tests and updates the associated tests to use the valid
xpath selectors (in Donations and Purchases).

This doesn't change the behavior if for non-numeric line item
quantities: it still silently discards that line item. If we want to
change that, we might want a client-side validation/to play with the
field type.