Fix tests broken due to zero-length window
Without this, the range is based on a date, without a time. Assuming both dates are the same, for example Jan 1, 1970, when sent to the database, this means that the `start_time` ends up being `Jan 1 1970 00:00:00` and the `end_time` also ends up being `Jan 1 1970 00:00:00`.  The user's intent here is to see all items matching for a given day, so make sure that the `end_time` is set to `23:59:59` and all relevant items are found.