Refactor into poro to make intent clearer
Moving this from a struct to a poro creates a very well-defined name, and we could use this in the future to add validations around the business logic for request items. Also this addresses a warning that would appear every time the method is called (in tests, as well as every time the controller action was hit).