WIP: shared example specs
In this spec we're signing in the user. Then we create 100 instances
of the respective model, for example the Adjustments model.

Knowing that the defaul per page set by Kaminara is 25, we can expect
the results to be returned across 4 pages -- with 25 on each page.

Then to confirm that the user is able to jump to another page, we use
capybara's visit command which kind of simulates navigation. We jump to
page 3.

Thanks to Kaminari's paginate helper method, we know that there are navigation buttons
on the page labled 1, 2, 3 and so on. Also, by running the app locally, we know
that when a user is on a particular page, that the corresponding navigation button
has a css "active" class on it.

So to confirm that we've successfully jumped to page 3, we search the page for
the link that has an "active" css class and then we confirm that link's text is 3.