RSpec shared example for pagination (#1401)
* WIP: Getting started with RSPEC shared example

* WIP

* WIP

* Trying to get tests to pass

* WIP

* WIP: shared example specs

In this spec we're signing in the user. Then we create 100 instances
of the respective model, for example the Adjustments model.

Knowing that the defaul per page set by Kaminara is 25, we can expect
the results to be returned across 4 pages -- with 25 on each page.

Then to confirm that the user is able to jump to another page, we use
capybara's visit command which kind of simulates navigation. We jump to
page 3.

Thanks to Kaminari's paginate helper method, we know that there are navigation buttons
on the page labled 1, 2, 3 and so on. Also, by running the app locally, we know
that when a user is on a particular page, that the corresponding navigation button
has a css "active" class on it.

So to confirm that we've successfully jumped to page 3, we search the page for
the link that has an "active" css class and then we confirm that link's text is 3.

* Added donations to shared example and cleaned up code.

* Adding co-authors

Co-authored-by: Eric Tillberg <eric.tillberg@gmail.com>
Co-authored-by: @lauramosher

* Fixing lint errors

* migrated db

* Revert "migrated db"

This reverts commit 71abc770862d24c4bcf631d0a4e8aa117fbbf02f.

* updating spec

* Tiny update to spec just to get a new build going.

* Fix

The build kept failing on this code because I set the Kaminari
config to 25 in the spec and it wasn't resetting to it's default.

In the hopes of getting the build to pass, we'll just rely on the
default being set in the kaminari config already.

* Addressing lint error

Co-authored-by: Eric Tillberg <eric.tillberg@gmail.com>