Add strong_migrations gem
**Why**
It can detect unsafe migrations in development so that you don't end
up with downtime in production. It also shows you how to perform the
migration in a safe manner.

This is one of the first gems I install in every Rails app that uses
migrations. It's a solid gem made by Andrew Kane from Instacart, who
has a track record of building helpful and well-maintained gems.