Adds line_items_total column to entities using itemizable
This is important: Since Itemizable is being used in multiple
models, we need to have counter caches in all of them. Of course
we could conditionally use the counter caches, but it is my opinion
that using this module would require further modifications in order
to keep the queries and views performant.

This will raise an exception requiring the column to be present if
using Itemizable in a new model. This is intended.