Extra wait time for locally flaky alert test
Bumps up the built-in default of 2 seconds of wait time.

When running locally the alert trigger sometimes has to wait for code to
be loaded before the rest of the dialog flow renders fully in the UI.
Particularly if this is one of the first tests to run. In CI code will
be eager loaded so this should not be an issue but in local envs waiting
a little longer means the test is much more likely to pass even if it is
the first (or only) test being run.