Eager loaded ci and alert spec stability (#2503)
* Eager load code for CI test runs

This is done primarily because it will improve the stability of a few
tests. Some system test cases will hit the default Capybara timeout if
they run early in the test suite in CI causing a failure purely because
the application is eager loading things behind the scenes. This change
should improve the stability of those tests on CI however they still may
be updated separately for improved stability in local testing.

Per the GitHub Actions docs tests have a CI=true flag by default:
https://docs.github.com/en/actions/reference/environment-variables#default-environment-variables

Typicaly you want to keep eager loading off to really quickly execute
individual tests.

In the CI environment a large batch of tests are run all at once so the
entire codebase typically needs to be loaded anyway. Loading it up front
should not negatively impact the class loading time and in some cases
might even speed things up a bit.

Eager loading is disabled locally because running a single spec takes
approximately 2-3 seconds longer to finish with eager loading enabled.
Running a single test shows even more time spent purely to loading code
that is irrelevant to the test for most test cases.

Per the simplecov docs the eager loaded mode can also produce more
accurate coverage when running tests with spring enabled:
https://github.com/simplecov-ruby/simplecov#want-to-use-spring-with-simplecov

* Extra wait time for locally flaky alert test

Bumps up the built-in default of 2 seconds of wait time.

When running locally the alert trigger sometimes has to wait for code to
be loaded before the rest of the dialog flow renders fully in the UI.
Particularly if this is one of the first tests to run. In CI code will
be eager loaded so this should not be an issue but in local envs waiting
a little longer means the test is much more likely to pass even if it is
the first (or only) test being run.