Put wrapper container around simple forms
Because edit forms run the loop once-per-item being listed, we end up with many containers with the same ID and each item in the list is in its own wrapper. When adding a new item, this means the first wrapper is found, with the single first item, and the new item is added inside the first wrapper. This change puts the wrapper around the simple_fields_for call so that items are properly added to the bottom. This is how purchases and donations are already setup in their similar views