Enables Display of Omitted Items in Adjustments
Closes #2543

This was a little tricky.

The inventory_storage_location_path action was used by a few different
places, so I parameterized the variant result.

The result was piped through a JSON builder using inventory item
records. This was pretty easy to Duck type a Struct in place and pass
that in alongside the existing inventory items. I added spec coverage
for both the feature overall as well as some request specs for the route
and its variants (which hadn't existed previously).