Revert "Try fixing the time on the machine level"
This reverts commit 87039ac (PR #2674)

It was needed because example groups in the Dashboard spec created
a single set of records using relative dates
(e.g., #beginning_of_week, #beginning_of_year, etc.) and made
assumptions about which records would/not be in each of the various
date ranges--assumptions which were faulty on various dates.

It is no longer needed because examples for each time range in the
Dashboard spec now generates records with timestamps relevant to the
time range "under test".