Begin converting Donation sys. spec to page obj.
Add OrganizationDonationsPage (just enough for examples so far) and
change the first few examples to use it.

Moved "When visiting the index page" block's contents out a layer
because I don't feel that title really adds any value.

Removed the "Displays Total quantity on the index page" because it
doesn't seem to be confirming anything useful. When I forced a failure,
the screenshot showed 2 0-item, $0-value donations and the only "20"s
in the table were the donation dates.

Looks like that example was added in 6c4e071. I rolled back to that
commit to get more context, but bundle install failed because a
specified gem version is no longer available.

I also split the "With an inactive item in a donation" to a context
block in hopes of increasing clarity.

Seeing nothing using the Donations created in the before block, I
removed the block.