Update Mailers to use the Default Mailer; Update Default Mailer content (#2546)
* Update Mailers to use the Default Mailer; Update Default Mailer content

Closes #2517

Two things are happening here:

The default mailer has been updated with respect to the new HTML
provided in #2517.

Some of our mailers were using full HTML documents rather than deferring
to the default mailer. Two of them (partner_mailer and
request_cancel_partner_notification) were complex and used full HTML
documents that had style tags and everything. I don't know how much of
that was necessary so I modified the default mailer to look for two
content blocks (head and body) and render them separately. The CSS is
now rendered to the :head block, if it exists. If no :head exists, it
just yields it normally.

I have not been able to test this. I did actions that I thought would
have triggered mails but there appears to be additional configuration on
the environment that I don't have set up :shrug: -- please test this and
ensure that the emails are rendering like they should be.

* Temp fix the broken mailer spec

* Remove temp fix

* Update layout to match attached file

Co-authored-by: Edwin <edwinthinks@gmail.com>
Co-authored-by: Daniel Orner <daniel.orner@wishabi.com>