Stub send_notification in two Distribution system specs to stop errors
Fixes #3161

In the CI environment only, two of our system specs were failing because
of an organization logo attachment issue. This is an unintended side
effect that is not within the test path of the features being tested by
the examples, so we can remedy this by just removing it.

I tested this by setting the DistributionPdf class to raise an error
when the :logo_image was set. This allowed me to know for certain when
the side effect was bypassed.

The remedy was to do a trivial stubbing of :send_notification, so that
it just doesn't even go down that path while testing the other stuff. It
correctly resolved the failures in my testing environment.