Adds filtering on distributions by item_id
Closes #3070

When filtering by an item type, it now displays all distributions that
contain that item and however many of that item specifically are
contained in the distribution.

The heading is now updated to display that quantity.

Some caveats:

  - the implementation is KLUNKY. It's using a basic if/else for both
    the table heading and also the table data
  - the helper being used is hyper-specific to quantity by item_id. This
    could maybe generalized later. or not. whatever you feel like :)
  - I did test it with a non-existent item ID and it gracefully
    displayed nothing

I modified the existing filter system spec to also verify that the TH
and TD tags both display the new text, correctly.

While fixing this, I identified a bug in the date-range params. I will
push up a separate PR for that.