Adds distribution filtering by item_category
Closes #3069

Similar to #3070.

The previous :quantities_by_category was actually using vestigial code
from the previous pass at categories (that was later stripped out, and
then later re-visioned). The new method uses the current ItemCategory
model as a basis.

I did my best to avoid N+1s but it's possible the queries could be
improved further, and more computation done in SQL.

The additional filtering field _does_ bump the the fields down a bit in
the UI, but that felt out of scope for this issue (no guidance was
provided on how to address the UI) -- I presume someone will have
opinions on how to address that separately. :)

I tested this both with a system spec, and also manually: I verified
that by filtering a specific item category, it will show the correct
quantity for each entry. The heading is also set correctly.

Exporting needs to be tweaked but I'll add that separately.