Reverts previous changes and applies a fix that directly fixes the cause
Thanks to @awwaiid for the spelunking!

The problem was that the org's logo was being purged off, and since Org
was persisted to improve test performance, this change was persisting
also, leading to flaky behavior.

This change stubs the method to see the logo, instead of actually
deleting it, which allows the spec to simulate the logo being missing
rather than actually making it missing.