Modifies ExportDistributionsCSVService to get @distributions directly
Previously, this was receiving a collection of ids from the controller,
and then re-running the query. This would absolutely be the correct way
to do it if/when this export is running from an async job (where we can
only use primitives). With the current state, however, it results in
running a semi-complex query twice.

We're only using the service object in two places, so the change has a
small footprint.

I also added an optional "filters" parameter and passed in
:filter_params to it on both controllers. This will be useful later when
applying filters to the export set to get it to mirror what is seen on
the index view.