Stub send_notification in two Distribution system specs to stop errors (#3220)
* Stub send_notification in two Distribution system specs to stop errors

Fixes #3161

In the CI environment only, two of our system specs were failing because
of an organization logo attachment issue. This is an unintended side
effect that is not within the test path of the features being tested by
the examples, so we can remedy this by just removing it.

I tested this by setting the DistributionPdf class to raise an error
when the :logo_image was set. This allowed me to know for certain when
the side effect was bypassed.

The remedy was to do a trivial stubbing of :send_notification, so that
it just doesn't even go down that path while testing the other stuff. It
correctly resolved the failures in my testing environment.

* Rubocop fixes

* Stubs the logo for PDFs so it always uses the default logo in these two request specs.

* Reverts previous changes and applies a fix that directly fixes the cause

Thanks to @awwaiid for the spelunking!

The problem was that the org's logo was being purged off, and since Org
was persisted to improve test performance, this change was persisting
also, leading to flaky behavior.

This change stubs the method to see the logo, instead of actually
deleting it, which allows the spec to simulate the logo being missing
rather than actually making it missing.

* Removes focus. oops