Adds distribution filtering by item_category (#3224)
* Adds filtering on distributions by item_id

Closes #3070

When filtering by an item type, it now displays all distributions that
contain that item and however many of that item specifically are
contained in the distribution.

The heading is now updated to display that quantity.

Some caveats:

  - the implementation is KLUNKY. It's using a basic if/else for both
    the table heading and also the table data
  - the helper being used is hyper-specific to quantity by item_id. This
    could maybe generalized later. or not. whatever you feel like :)
  - I did test it with a non-existent item ID and it gracefully
    displayed nothing

I modified the existing filter system spec to also verify that the TH
and TD tags both display the new text, correctly.

While fixing this, I identified a bug in the date-range params. I will
push up a separate PR for that.

* WIP

* Adds distribution filtering by item_category

Closes #3069

Similar to #3070.

The previous :quantities_by_category was actually using vestigial code
from the previous pass at categories (that was later stripped out, and
then later re-visioned). The new method uses the current ItemCategory
model as a basis.

I did my best to avoid N+1s but it's possible the queries could be
improved further, and more computation done in SQL.

The additional filtering field _does_ bump the the fields down a bit in
the UI, but that felt out of scope for this issue (no guidance was
provided on how to address the UI) -- I presume someone will have
opinions on how to address that separately. :)

I tested this both with a system spec, and also manually: I verified
that by filtering a specific item category, it will show the correct
quantity for each entry. The heading is also set correctly.

Exporting needs to be tweaked but I'll add that separately.

* Rubocop fixes