BUGFIX: date range params on distribution was using label for value (#3222)
* BUGFIX: date range params on distribution was using label for value

I found this bug while on a different feature.

The partial for app/views/shared/_date_range_picker.html.erb was using
@selected_date_range_label for the second parameter of the
text_field_tag. This is the value used for the actual value of the
field, and is not what the form is expecting. (The label is stored in a
hidden field tag, presumably to be used by other stuff).

To reproduce the error, normally:

  1. Go to distributions#index
  2. Apply a filter (to seed filter_params); note the new value in the
     date-range field
  3. Click "Filter" again, you should see an error

By changing the field to use the :date_range_params helper method
(already existing), this restores the behavior to what is expected.

I am not able to consistently get the helper UI JS thingy to display,
but it does display sometimes? The lack of displaying is _probably_
unrelated? I'm not seeing JS errors show up in the console. I'm not sure
how it's wired up.

* Reverts _date_range_picker change since this is addressed instead by the controller change