Fixed the Issue
[Fixed] Name does not show up

[Fixed] The field for other agency type doesn't show up

[Fixed] Currently Providing diapers, Case Management, Evidence Based, Storage Space, Income Requirement and Income Verification show up as true/false instead of Yes/No

[Fixed] Program Name(s), Descriptions(s) and Age are missing a colon

[Changed] Currently Provide diapers -> Currently Providing Diapers

[Fixed] Child-based requests, Requests for individuals, and Quantity-based requests all show up in the Agency Stability section when they are already in the Settings section

[Changed] Additional documents and Settings pieces are now arranged in the same format as the rest of the page (no longer bullet points)

-- Additional Considerations --

Additional documents field becomes blank if a document is not added every edit step, not sure why

"How Are Essentials (e.g. diapers, period supplies) Used In Your Program?" And "Do You Receive Essentials from Other Sources?" Could be re-worded to be consistent.

"More docs Required" is a text box, should it be a check box?

Income Requirement was labeled as "Income Requirement Description" but it is a check box. Should the income requirement be a number instead of a boolean?

Percentages are all integers, should they be changed to floats?

Do You Track Program Results (in editing) -> Evidence Based?

Distributor type appears to be an unused field