remove dead code - unused helper `after_sign_in_path_for`.
It was supposed to override devise's method `after_sign_in_path_for`. However, to make it acutally wotk, it would need to be defined directly in ApplicationController or the AppliactionHelper would need to be directly included in the ApplicationController.

So in that form, it was only available through `helpers.after_sign_in_path_for` and it was never invoked.