Distributions summary
Fixes issue #3723

Changes made:
This creates a new area for distributions summary to live under audits, adds system, page, and request spec for testing.

sidebar - we also fixed the weird issue with the sidebar for the annual reports, distributions by county, and itemized donations & distributions that were not staying selected.