feat: Item Inventory for Banks with many locations (#3784)
* Add new table with collapsible locations for banks with many locations

The "Items, Quantity, and Location" table has a nasty horizontal scroll
if the bank has many locations, which can be common for banks that allow
volunteers to take inventory back to their residences to package into
kits. This is very unwieldly for them, but other banks with few
locations have a great scannable view of where their inventory is. We're
reproducing this table with the locations collapsed under a heading for
banks with many locations in a new tab to preserve the experience for
banks with few locations.

ExpandableTable from AdminLTE is being used to handle the expansion.
Location is being stored slightly differently in the query to make it
easier to access.

* Add new styling for expandable tabes to hint that they can be expanded when clicked.

* Location name could be duplicated, so use id to identify the location instead.

Resolves #3776