Fix active styling for Reporting & Auditing in the sidebar.
Reporting & Auditing was not expanding for most of the navigation items
in the list. In addition, the currently selected report was not
highlighted.

active_class and menu_open? now check if the listed path is in the
controller_path instead of doing an exact match on controller_path.

In addition, the report nav items have been updated to reference their
own path, so all reports aren't listed as active when one is selected.