fix: renames Itemizable.to_a to line_item_hashes (#3909)
Itemizable has a method to_a which returns an array of hashes
representing the line items in that itemizable object. to_a
is meant to convert an object to an array - instead, this
returns an array of child members of the record.

This renames the method so that when we work with the line items
inside an Itemizable, we explicitly use a array of hashes
rather than the parent Itemizable itself.

The main methods affected are by this change are
StorageLocation.increase/decrease. These methods were overloaded
accepting both Itemizable objects and arrays of hashes. Now, they
will exclusively accept arrays of hashes.

The Itemizable.to_a method is not fully removed,
it currently logs an error and calls the correct method.

Fixes #3909