fix: individuals request preserve values
Fixes #4135

When individuals requests were submitted with incorrect values
the error page did not preserve correct or incorrect inputs. This
was happening because the controller created a new request on
error.

Now the new request will be populated with the items from the
failed request.

I attempted to reuse the failed request, however, because the
individuals request is a wrapper around a family request that
behavior was not possible.