feat: serve last user role on login
When a user signs in they will now be served their most recently used
role in priority order: role from session > most recent role from
database > most permissive role they have.

Created a foreign_key key from users to user_roles to track the last
used role. The foreign_key is in addition to a has_one :last_role from
the user_role to the user with dependent :nullify so if the role is
revoked the last_role is also revoked.

The behavior of finding the role was move to UsersRole class.

Co-authored-by: Veez Remsik <almighty@veez.us>