Fix issue with hidden 'Donation Site' field not being shown on form submission
- The issue occurred when creating a donation and selecting 'Donation Site'
as the 'Source' without choosing a value for the 'Donation Site' field.
After submitting the form, the error message was displayed correctly,
but the 'Donation Site' field remained hidden.
- Removed the `hide_fields_with_no_value(fields_to_hide)` and renamed function
on line 64 `handleSourceSelection`. This function is called on page load and
on the 'change' event, and checks if a value is selected in the Source field and
toggles the display of connected fields.