refactor: remove more test order dependence (#4273)
* fix: remove more test order dependence

As discussed here https://github.com/rubyforgood/human-essentials/pull/4264,
some tests modify global variables and expect them to
be reset before each test. This makes it more difficult
to move forward with removal of pre-seeding.

This commit refactors certain tests to be less brittle to
changes in global state and to rely less on assumptions
on global state. Ex: not assuming that there are 0 records
of a certain type when a test starts, etc.

* fix: minor cleanup / lint