feat: replace org_user role deactivate with removal
The issue was that banks could "deactivate" a user
(which is essentially a soft delete of the whole User record).
However, if that user then went to a Partner they would not be able to invite
that user because that user would exist (new User cannot be created)
but be soft deleted (existing User is treated like they didn't have permissions).

We changed the logic around deactivation to address the role not the user.

I attempted to implement logic around deactivating roles themselves
in #4392, #4382. But the logic became very complex.

We decided (#3587 (comment)) to go for a more architerurally simple solution
of removing the deactivation functionality and just deleting the roles.

To migrate the existing data I decided to reactivate any discarded user
with the org user role and remove the org user role from them. (My guess
is that those are the org users that actually got deactivated)