Adds input for the reminder schedule in the Edit Partner Group page
Builds an ActiveModel, ReminderSchedule, which takes the necessary information
and turns it into an IceCubeSchedule, which is what ultimate get saved in the db.

Builds the form for this reminder schedule. Still needs to conditionally show or hide
sections depending on if the user wants date or day of the week.