Filters out deactive partners on new distribution
This commit adds in a conditional check to the _form partial that
checks where the partial is coming from. If it is being rendered from
the 'new' view, it will then generate a partner list that filters out
the deactivated partners. If it is being rendered from anywhere else,
it will generate the complete partner list unfiltered.

[Ticket: 4515]