4515 - Deactivated partners should not appear in the dropdown when producing a new distribution. (#4639)
* Filters out deactive partners on new distribution

This commit adds in a conditional check to the _form partial that
checks where the partial is coming from. If it is being rendered from
the 'new' view, it will then generate a partner list that filters out
the deactivated partners. If it is being rendered from anywhere else,
it will generate the complete partner list unfiltered.

[Ticket: 4515]

* Updates the re-render of create on a fail to also filter out the deactivated partners

* Adds in unit tests for distribution dropdown

* Removed unneeded variable declaration for unit tests

* Moves the partner_list query from the view into
the controller actions

* Moved rspec tests from 'system' to 'requests'