4473 - fixes reminder day vs deadline display bug
There was a bug in which if the reminder day of the month was the same
as the deadline day, it would dislplay the error message even if the user
changed to the day of the week option. Now the error message is hidden if
the user is selecting the day of the week option.