Fix flaky distributions by county system spec
Distribution by county system specs are flaky when run at certain times
of day because the system time zone (for Capybara) and server time zone
(Rails) are in PST. But timestamps are stored in the database in UTC
time. Normally this isn't an issue because Active Record will convert
to UTC for queries. But since this is a raw SQL query we need to convert
to UTC.

This is not an issue for production though, because the timezone in
production is UTC.